shell.xml: Add org_kde_plasma_shell_surface interface
Just like xdg_shell we need an interface for surfaces.

A shell will create a org_kde_plasma_surface object when
a special window is created (such as desktop, notifications, ...)
and the compositor will be able to identify it as a shell window
rather than an application window (for which xdg_surface or another
protocol will be used).

Since org_kde_plasma_shell.set_position is moved to
org_kde_plasma_surface a new set_global_position request is added
to org_kde_plasma_shell.